broken down the gateway_client to multiple files
