+++ ibeam/config.py
@@ -3,6 +3,8 @@ import os
 import sys
 from pathlib import Path
 
+from ibeam.src import var
+
 initialized = False
 
 
@@ -15,6 +17,7 @@ def initialize():
     # logger.setLevel(logging.DEBUG)
     stream_handler = logging.StreamHandler()
     stream_handler.setFormatter(logging.Formatter('%(asctime)s|%(levelname)-.1s| %(message)s'))
+    logging.getLogger('ibeam').setLevel(getattr(logging, var.LOG_LEVEL))
     logger.addHandler(stream_handler)
 
     _this_filedir = os.path.abspath(os.path.dirname(__file__))
